feat(analytics): moving scroll tracking to dotcomshell
