+++ packages/services/src/services/Analytics/index.js
@@ -5,4 +5,4 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-export { default as RegisterAnalyticsEvent } from './Analytics';
+export { default as AnalyticsAPI } from './Analytics';
